Changes made to long method, parseTasks in Parser
