+++ src/main/java/seedu/duke/Launcher.java
@@ -1,6 +1,7 @@
 package seedu.duke;
 
 import javafx.application.Application;
+import seedu.duke.ui.Main;
 
 /**
  * A launcher class to workaround classpath issues.
